add a lot of some changes

+++ src/movie/movie.service.ts
@@ -1,5 +1,4 @@
-import { Injectable } from '@nestjs/common';
-import { DeleteResult } from 'typeorm';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { MovieRepository } from './movie.repository';
 import { FilesService } from '../files/files.service';
 import { Movie } from './movie.entity';
@@ -7,6 +6,7 @@ import { CreateMovieDto } from './dto/create-movie.dto';
 import { ActorsService } from '../actors/actors.service';
 import { DirectorsService } from '../directors/directors.service';
 import { UpdateMovieDto } from './dto/update-movie.dto';
+import { UserService } from '../users/user.service';
 import slugify from 'slugify';
 
 @Injectable()
@@ -16,6 +16,7 @@ export class MovieService {
     private readonly filesService: FilesService,
     private readonly actorsService: ActorsService,
     private readonly directorsService: DirectorsService,
+    private readonly userService: UserService,
   ) {}
 
   async findAll(page: number = 1): Promise<{ count: number; items: Movie[] }> {
@@ -34,8 +35,16 @@ export class MovieService {
     };
   }
 
-  async findOne(id: number): Promise<Movie> {
-    return this.movieRepository.findOne(id);
+  async getMovie(id: string): Promise<Movie> {
+    const foundMovie = this.movieRepository.findOne(id, {
+      relations: ['cast', 'directors'],
+    });
+
+    if (!foundMovie) {
+      throw new NotFoundException('Movie not found');
+    }
+
+    return foundMovie;
   }
 
   async create(movie: CreateMovieDto): Promise<Movie> {
@@ -80,12 +89,19 @@ export class MovieService {
     return this.movieRepository.updateOne(id, movieData, { directors, actors });
   }
 
-  async delete(id: string): Promise<DeleteResult> {
-    return await this.movieRepository.delete(id);
+  async delete(id: string): Promise<Movie> {
+    const movie = await this.getMovie(id);
+
+    if (!movie) {
+      throw new NotFoundException(`Movie with id = ${id} does not exist`);
+    }
+
+    await this.movieRepository.delete(id);
+    return movie;
   }
 
   async uploadFiles(
-    movieId: number,
+    movieId: string,
     preview: Express.Multer.File,
     videoFile: Express.Multer.File,
   ) {
@@ -101,6 +117,73 @@ export class MovieService {
     return {
       preview: previewModel,
       videoFile: videoFileModel,
+    };
+  }
+
+  async getSuggestedMovies() {
+    const newReleases = await this.movieRepository.find({
+      relations: ['cast', 'directors'],
+      take: 20,
+      where: {
+        isPublished: true,
+      },
+      order: {
+        releaseDate: 'DESC',
+      },
+    });
+
+    const popularOnKinobox = await this.movieRepository.find({
+      relations: ['cast', 'directors'],
+      take: 20,
+      where: {
+        isPublished: true,
+      },
+      order: {
+        id: 'ASC',
+      },
+    });
+
+    const trendingNow = await this.movieRepository.find({
+      relations: ['cast', 'directors'],
+      take: 20,
+      where: {
+        isPublished: true,
+      },
+      order: {
+        title: 'DESC',
+      },
+    });
+
+    return [
+      {
+        name: 'New Releases',
+        items: newReleases,
+      },
+      {
+        name: 'Popular on Kinobox',
+        items: popularOnKinobox,
+      },
+      {
+        name: 'Trending now',
+        items: trendingNow,
+      },
+    ];
+  }
+
+  async addMovieToList(userId: string, movieId: string) {
+    const movie = await this.movieRepository.findOne(movieId);
+    if (!movie) {
+      throw new NotFoundException('Movie not found');
+    }
+    return this.userService.addMovieToList(userId, movie);
+  }
+
+  async getList(userId: string) {
+    const { list } = await this.userService.findOneById(+userId, {
+      relations: ['list'],
+    });
+    return {
+      list,
     };
   }
 }
